Style fixes, include guard

+++ pixel.cpp
@@ -86,7 +86,7 @@ uint8_t PixelBone_Pixel::h2rgb(uint8_t v1, uint8_t v2, uint8_t hue) {
 //
 //   lightness:  0 to 100 - how light the color is, 100=white, 50=color, 0=black
 //
-uint32_t PixelBone_Pixel::HSB(uint8_t h, uint8_t s, uint8_t l) {
+uint32_t PixelBone_Pixel::HSB(uint16_t h, uint8_t s, uint8_t l) {
   uint8_t r, g, b;
   uint8_t v1, v2;
 
@@ -100,9 +100,9 @@ uint32_t PixelBone_Pixel::HSB(uint8_t h, uint8_t s, uint8_t l) {
   if (s == 0) {
     r = g = b = l * 255 / 100;
   } else {
-    if (l < 50 )
+    if (l < 50)
       v2 = l * (100 + s);
-    else 
+    else
       v2 = ((l + s) * 100) - (s * l);
 
     v1 = l * 200 - v2;
@@ -112,7 +112,6 @@ uint32_t PixelBone_Pixel::HSB(uint8_t h, uint8_t s, uint8_t l) {
   }
   return Color(r, g, b);
 }
-
 
 // Query color from previously-set pixel (returns packed 32-bit RGB value)
 uint32_t PixelBone_Pixel::getPixelColor(uint32_t n) const {
@@ -165,7 +164,7 @@ uint32_t PixelBone_Pixel::wait() {
 }
 
 void PixelBone_Pixel::clear() {
-  for ( int i = 0; i < num_pixels; i++ ) {
+  for (uint16_t i = 0; i < num_pixels; i++) {
     this->setPixelColor(i, 0, 0, 0);
   }
 }
